+++ lib/views/text_field_item_view.dart
@@ -2,49 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_tts/flutter_tts.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
-enum Language {
-  english,
-  japanese,
-}
-
-extension LanguageExtension on Language {
-  String get text {
-    switch (this) {
-      case Language.english:
-        return 'English';
-      case Language.japanese:
-        return '日本語';
-    }
-  }
-
-  String get translatorLabel {
-    switch (this) {
-      case Language.english:
-        return 'en';
-      case Language.japanese:
-        return 'ja';
-    }
-  }
-
-  String get speakingLabel {
-    switch (this) {
-      case Language.english:
-        return 'en-US';
-      case Language.japanese:
-        return 'ja-JP';
-    }
-  }
-
-  TextInputType get textInputType {
-    switch (this) {
-      case Language.japanese:
-        return TextInputType.text;
-      default:
-        return TextInputType.emailAddress;
-    }
-  }
-}
+import 'package:mobile_translation_app/models/language.dart';
 
 class TextFieldItemView extends StatelessWidget {
   const TextFieldItemView({
